Fix inconsistencies in tab bar visibility
